backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pymongo==4.3.3
python-dotenv==1.0.0
joblib==1.3.2
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
scipy==1.10.1
python-multipart==0.0.6